Balance overhaul #1.
Big nerfs and buffs.  I've clarified the vanilla testing and general
balance in a way that I think will be conducive to a strong metagame
with a lot of variety.  For the existing cards I had to rework some
effects and a ton of stats; for the new ones I'm just going to be more
mindful of the ideal balance.